Items that share both the same slot and the same equipslot (hands/body/head) will now be equipped to make space for an incoming item.
For example, when both Axe and Spear are set to slot #2 and a currently equipped Axe is unequipped when the Spear occupies slot #2, the Spear will now be equipped to make space,
rather than being moved to some other slot. This is most likely what the user intended when saving 2 pieces of equipment in the same slot. In addition, this is also what happens already when
the Spear would be equipped from slot #2, in that case the Axe would move back to slot #2 as well, so adding this functionality means the behavior symmetrical and it does not matter if you equip the Spear
or unequip the Axe, whereas previously the behavior would be different.