Equipping an item to make space now also occurs for items of the same prefab, regardless of remaining uses.
For example, this means that unequipping an Axe when you have another Axe in the saved slot will now equip the other Axe rather than move it away.
This is configurable through the "Allow Equip For Space" option. When disabled, this mod will never equip an item to make space.
Also removed backticks in comments and some other minor improvements.